:recycle: refactor Layout
