--- src/components/BlogLayout.tsx
@@ -1,14 +1,0 @@
-import React, { FC } from "react";
-import { Box } from "rebass/styled-components";
-import RouteLink from "./common/RouteLink";
-
-const BlogLayout: FC = ({ children }) => {
-  return (
-    <Box mx={[3, 4, 5]}>
-      <RouteLink href="/">Home</RouteLink>
-      <Box>{children}</Box>
-    </Box>
-  );
-};
-
-export default BlogLayout;
